refactor routing_a8w4 name to unify dsv4 and grouped topk + sigmoid
